requirements.txt
coverage==7.6.4
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
